fix: removed modules due to eslint and browser errors
