+++ index.html
@@ -7,7 +7,7 @@
   <title>Keyboard</title>
   <link rel="icon" href="src/img/keyboard.ico">
   <link rel="stylesheet" href="src/style.css">
-  <script type="module" src="src/script.js" defer></script>
+  <script src="src/script.js" defer></script>
 </head>
 <body>
