+++ src/style.css
@@ -1,4 +1,4 @@
-.container{
+.container {
   max-width: 1200px;
   margin: 0 auto;
 }
@@ -11,6 +11,7 @@
   display: block;
   margin: 20px auto;
 }
+
 .key {
   min-width: 40px;
   height: 40px;
@@ -20,27 +21,42 @@
   text-transform: uppercase;
   padding: 10px;
   cursor: pointer;
+  transition: all 0.7s;
+  box-shadow: 0px 2px 5px 5px black;
 }
+
 .big-key {
-  /* min-width: 80px; */
   text-transform: capitalize;
 }
 
 .keyboard {
   display: grid;
-  grid-template-columns: repeat(15,[col]1fr);
-  grid-template-rows: repeat(5,[row]1fr);
+  grid-template-columns: repeat(15, [col]1fr);
+  grid-template-rows: repeat(5, [row]1fr);
   grid-gap: 5px;
+  background-color: black;
+  border-radius: 5px;
+  padding: 8px;
 }
-.backspace, .tab, .caps , .enter, .shift, .control{
+
+.backspace,
+.tab,
+.caps,
+.enter,
+.shift,
+.control {
   grid-column: span 2;
 }
+
 .space {
   grid-column: span 5;
 }
-.key:hover{
+
+.key:hover {
   background-color: rgb(189, 248, 248);
 }
+
 .active {
   background-color: aqua;
+  box-shadow: 0px 2px 5px 5px azure;
 }
